+++ src/renderer/store.js
@@ -1,4 +1,5 @@
 import { writable } from 'svelte/store';
 
-export const url = writable(localStorage.getItem('url') || 'https://m.youtube.com');
+export const url = writable(localStorage.getItem('url') ||
+  'https://m.youtube.com');
 export const webview = writable({});