+++ CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ---
 
+### Changed
+
+- Eslint config.
+- Jest config.
+
+### Removed
+
+- Removed the prettier.
+
+
+---
+
 ## [1.5.2] - 2020-06-23
 
 ### Added